requirements.txt
asgiref==3.7.2
boto3==1.26.156
botocore==1.29.156
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
DateTime==5.1
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.1.0
django-mysql==4.10.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.1.12
et-xmlfile==1.1.0
hvac==1.1.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.7.0
MarkupSafe==2.1.3
mysql-client==0.0.1
numpy==1.25.1
openpyxl==3.1.2
pandas==2.0.3
pycparser==2.21
pyhcl==0.4.4
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
validators==0.20.0
win32-setctime==1.1.0
zope.interface==6.0
